Fixes bug when consecutively loading two queries of the same type
The application was not working properly when a query result was loaded
and a result of the same type was already displayed. The problem was
that the table was not updated.

This commit fixes this error, and now the tables are aware of result
changes and, when a result is changed, the table is updated.