Uses the work name when downloading Jaccard and Cmap CSV results
This commit sets the name of the downloaded CSV files containing the Jaccard and Cmap
results to the name of the corresponding works. If the works have no name, the name of
the file is still "Jaccard" or "Cmap" followed by the work ID.

This commit also uses the work ID in the works list when works have no name and
displays an error message when a query fails.