Changes cell type and subtype filters behaviour
This commit changes the behaviour of the cell type and subtype filters when
retrieving predictions (interactions) or signatures (in jaccard queries).

The motivation of this change is allowing users to introduce first a cell type
(or subtype), named type 1, and look for signatures that has either type A or
type B equal to this value. If a second type (or subtype) is specified, then
signatures that has either type A equal to type 1 and type B equal to type 2 or
signatures that has either type A equal to type 2 and type B equal to type 1
are retrieved.

Accordingly, the resources to list the possible cell type (or subtype) 1 or 2
have been modified. By one hand, a request of cell type (or subtype) 1 values
returns all distinct values in both type (or subtype) A and B. On the other
hand, a request of cell type (or subtype) 2 values, which always require the
cellType1 (or cellSubType1) parameter, will return all cell type (or subtype)
values that appear either in cell type (or subtype) A or B with the specified
cellType1 (or cellSubType1).