Modifies database filters behaviour
This commit modifies the database filters behaviour so that:
- Now, there are three basic filters and the complete set of filters are under a
collapsible panel called "Advanced Filters". The three basic filters also appear
in this panel of advanced filters. User should select at least the organism and the
cell type and subtype 1 to filter the database when the basic filters are used.
- The cell type and cell subtype filters now appear in a single selection component.
This change also affects to the Signature Comparison (Jaccard) configuration panel.