Adds value validation and clearable option to NumberFieldComponent
This commit adds value validation to the NumberFieldComponent so that it does not
emmit a value change event if the value is not in the specified [min, max] range
of the component. Moreover, if the value is not valid, the background color of the
input is changed to red.

This commit also adds a clearable input option to allow hidding the clear button.