Adds query parameters to the database URL
This commit makes the neccessary changes to manage query parameters in the database
URL:
- When database filters are changed they are added as query parameters
to the current URL.
- When the database is loaded, the URL is processed to obtain the query parameters.
These parameters are used to initialize the database filters and make the initial
data loading.

This commit also fixes one bug: the filters component did not emmit a change when
the max down FDR filter was changed.