Fixes overlapping points in graph views
This commit fixes an issue with overlapping points in graph views by adding all of
them to the tooltip.

To do this, a map is needed to know when in there are overlapping points in the (x,y)
position for which the tooltip is generated. Thus, when interactions are being mapped
to points to create the series, all interactions are added to a list at
"map.get(x+y)". Once the series have been created, all map entries with length equals
to 1 are removed since we are only interested in those with overlapping points.