Changes work results visualization
This commit changes the behaviour of the ResultsView component. So far, this
component aggregated the specific visualization components (Jaccard, CmapUpDown or
CmapGeneSet) depending on the type of the work under consideration. Now, when the
work finishes, this component redirects to these components by using navigation.

This way, the corresponding navigation menu is highlighted according to the section
the result belongs to when it the view is rendered.