Redesigns drug charts presentation
The drug charts are now presented in one single group instead of
using two groups (one for the prioritization plot and another one
for the pie plots).

As a result of this change, the *SignatureResultsSummary components
have been refactorized and merged into one single component. To
implement this component the CmapResultsSummary has been reused,
replacing its code with the newly unified one.