Adds a new component to display the signature information
The signature information in tables must be displayed as the signature name followed
by three links to PubMed, the source DB URL, and a DREIMT query to view it. Thus, a
new component to display has been created and it is now used in both the main
database table and the Jaccard results view.