Fixes a bug when generating the signature summary
This commit fixes a bug generating the signature summary that caused that in some
cases the treatmentA and treatmentB were incorrect.

The SignaturesSummaryHelper uses two maps to store the treatmentA and treatmentB
strings using the signature names as keys in order to avoid processing the same
treatment lists several times for the same signature. However, these maps should be
passed to the "_getExplanation" as parameter instead of allowing this method to access
them (as previously). This way, they can be also interchanged when the method is
invoked, otherwise in some cases the value associated to a signature name is taken
from the incorrect map.