Adds cell type effect filter and reformats filter fields
The cell type effect filter has been added to the database query
panel. This filter allows to select whether a boost or inhibit
effect is produced. In addition, the filter field design has been
updated for a better presentation.

Also, filter fields labels have been rewritten so that only the
first letter is in uppercase. To keep the entire interface
consisten with this rule, other texts have been modified to
meet it (e.g. table headers, section titles, and so on).