Fixes bad visualization of highlighted text in autocomplete
A weird bug was causing that highlighted text in autocomplete lists
(as in the FilterFieldComponent) was displayed as it did not have
spaces around. This caused a bad presentation of the item list.

This commit fixes this error by using a "display: contents" option in
the HTML element used to highlight the text.