Modifies the Cmap query to introduce case and reference types
This commit modifies the Cmap query panel to allow users to introduce case and
reference types. If not provided, a default value is assigned to them when making the
REST query since the backend needs these parameters.

Since these parameters are optional, a new field has been introduced in the panel
to allow users to specify the input query type (signature, signature up, signature
down, or geneset). This information is used to validate the input genesets (e.g.
a signature down query only requires down genes) and make the apropriate query.