Add cell type or subtype filters
This commit introduces support for sending queries with the new type or subtype
filters. So far, when the user typed a cell type/subtype filter the format was
"type / subtype", so that the cellType1 and cellSubType1 (or 2) filters were sent in
the queries. Now, if a word without a "/" separator is introduced, this is send in
the new cellTypeOrSubType1 (or 2) filter, in order to allow the word to be in either
the type or the subtype. If the separator is introduced (manually or by clicking in
the combobox options), then the cellType1 and cellSubType1 (or 2) filters are
created.