Modifies filter fields behavior
The filter field component has been changed to modify its behavior.
Now this component receives a model with the logic to retrieve the
valid values for the field and with the filter value. With this
model, the component now has more control on when the valid options
must be updated, reducing the calls to the backend and improving
response speed.