Disables database parameters when loading data
The database view now disables the query parameter fields while the
data is loading to avoid parameter changes while a request is on
course.

In addition, a loading spinner is now also shown in the query
parameters panel, as the one shown in the table pagination panel is
not visible in queries with lots of results.